requirements.txt
flask
flask_restful
flask_limiter
flask_pymongo
flask_cors
pymongo
unidecode
dms2dec
requests
gunicorn
python-dotenv
dnspython
cryptography